refactor: rename Paciente related classes
